fix: cast id to `signed long long' for ffsll
__builtin_ffsll takes a signed long long, not a uint64_t. Explicitly
cast to `signed long long' before passing it.

Caught with clang tidy.

+++ src/nccl_ofi_idpool.c
@@ -108,7 +108,7 @@ int nccl_ofi_idpool_allocate_id(nccl_ofi_idpool_t *idpool)
 	int entry_index = 0;
 	int id = -1;
 	for (size_t i = 0; i < num_long_elements; i++) {
-		entry_index = __builtin_ffsll(idpool->ids[i]);
+		entry_index = __builtin_ffsll((signed long long)idpool->ids[i]);
 		if (0 != entry_index) {
 			/* Found one available ID */
